Added Include and Exclude to packing
Can now include or exclude specific files in the pack step.

Including a file will search for that file's dependencies, which I'm not sure should be default behavior.

Fixed a glitch with paramters that could be added more than once where it would be linked to another paramter.

Some small refactorings