Added new FileInput control
The FileInput control is a textbox with a button that opens the file browser. This was originally a part of the Custom Program screen, but it's now separate and can be used anywhere.

Added FileInput controls to Include and Exclude parameters.

Added new properties to ConfigItem, ValueIsFile and Value2IsFile.
If these values are true, it will replace the value textbox with a FileInput control

Removed debug console logging in Include and Exclude parameters.