Fixed Particle File Packing
Fixed issue with not stripping "custom/blah/" from custom/blah/particles/p.pcf.

Fixed issue with packing particle manifest files that weren't generated with compilepal when we are generating one.

Fixed issue with generated particle manifest not having a valid internal directory by switching from replace to substring. Not sure why replace didn't work, but it wasn't removing the baseDirectory and substring is more correct as it removes from the beginning of the path instead of possibly from the middle.