Improved keyvalue parsing in vmtPathParser
Made parsing better by using regex so values with // in the middle don't get mangled.

This is not the best solution though, this project needs to use a real vdf parser in every part of the code that touches vdf files, but I do not have time for that.

ported fix for packing multiple particle manifests

Fixed case problems in manifest path packing. Cleaned up dupe manifest packing

Switched replace with substring when removing basedirectory from the filepath. Replace was failing because the 2 strings had different cases.

Cleaned up the dupe manifest packing fix.

more parsing fixes

send the full line to a new function. quotes are needed to extract values with spaces in them